pyproject.toml
[project]
name = "macmocker"
version = "1.0.0"
description = "macOS device testing suite"
authors = [
    {name = "David Crosby", email = "bing@bing-bong.co.uk"}
]
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[project.urls]
Repository = "https://github.com/david-crosby/macmocker"
